+++ backend/src/main/java/turip/content/controller/dto/response/ContentPlaceDetailResponse.java
@@ -1,0 +1,23 @@
+package turip.content.controller.dto.response;
+
+import java.util.List;
+
+public record ContentPlaceDetailResponse(
+        TripDurationResponse tripDuration,
+        int contentPlaceCount,
+        List<ContentPlaceResponse> contentPlaces
+) {
+
+    public static ContentPlaceDetailResponse of(
+            int nights,
+            int days,
+            int contentPlaceCount,
+            List<ContentPlaceResponse> contentPlaces
+    ) {
+        return new ContentPlaceDetailResponse(
+                TripDurationResponse.of(nights, days),
+                contentPlaceCount,
+                contentPlaces
+        );
+    }
+}
