+++ backend/src/main/java/turip/content/service/ContentPlaceService.java
@@ -4,8 +4,8 @@ import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import turip.common.exception.custom.NotFoundException;
-import turip.content.controller.dto.response.MyFavoriteContentsResponse.ContentPlaceDetailResponse;
-import turip.content.controller.dto.response.MyFavoriteContentsResponse.ContentPlaceResponse;
+import turip.content.controller.dto.response.ContentPlaceDetailResponse;
+import turip.content.controller.dto.response.ContentPlaceResponse;
 import turip.content.domain.ContentPlace;
 import turip.content.repository.ContentPlaceRepository;
 import turip.content.repository.ContentRepository;
